Refactor API routes to use user-specific namespaces
- Moved all user-related API routes (`/beans`, `/brews`, `/analyze`, `/settings`, and `/images`) under `/api/users/*` for better organization and clarity.
- Updated client-side fetch calls to match the new API endpoints.
- Deleted redundant API modules (`analyze.ts`, `beans.ts`, `brews.ts`, `settings.ts`, and `images.ts`) and consolidated logic into user-specific routes.
- Adjusted `BrewContext` and `SettingsContext` to use updated API paths.
- Updated backend route definitions in `index.tsx` to align with the new API structure.

This refactor improves maintainability by grouping user-related resources under a consistent namespace.
